+++ CHANGES.rst
@@ -3,6 +3,8 @@ Changelog
 
 To be released
 --------------
+- Add deprecation warning for MonitorField. The default value will be `None`
+  instead of `django.utils.timezone.now` - when nullable and without a default.
 - Add Brazilian Portuguese translation (GH-#578)
 - Don't use `post_init` signal for initialize tracker